ci: ammend GitHub action by check, process & push
The GitHub action was ammended by a check to fetch and process the
remote README.md file only if the check sum of the principal branch
in the remote repository was changed.  This aims to shorten the
processing time, where suitable, when running once a day.  If there
was a change, the action should file and push a commit, too.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>