fix: provide non-iterative MWE to interface GitHub API
Previously, only Fortran projects with a fpm.toml file within a
principal branch named "master" qualified to enter the new
README.md file written.  However, git does not require the
principal branch (of deployment) to be named this way and many
projects hosted on GitHub use "main" as an alternative instead.
This commit provides

- the connection to the Github API
- an authentication for greater performance (up to 5k/h calls)
- the identification of principal branch, regardless of its name
- probe for presence of a README.md in the principal branch

Results are reported back to the CLI.  The earlier approach by
still is present in the file, however truncated .and. muted.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>