fix(automate_xfpm.yml): logic in if/fi block, reformat
Recognized by the automated daily check to the raw data, the logic
retaining a copy of most recent git hash of the raw data no longer
was functional.  This then caused to process the raw data even if
they were not edited in first place.  This error is addressed, the
yml file equally was lightly reformatted, too.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>